Pull out generic methods on helpdesk
